move plots to subpackage
