+++ geoglows/_plots/__init__.py
@@ -1,0 +1,4 @@
+from .plots import *
+
+
+__all__ = []
