+++ geoglows/_plots/matplotlib_retrospective.py
@@ -1,0 +1,34 @@
+import pandas as pd
+
+
+__all__ = [
+    'retrospective',
+    'daily_averages',
+    'monthly_averages',
+    'annual_averages',
+    'daily_variance',
+    'flow_duration_curve',
+]
+
+def retrospective(df, *, rp_df: pd.DataFrame = None, plot_titles: list = None):
+    return
+
+
+def daily_averages(df, *, plot_titles: list = None):
+    return
+
+
+def monthly_averages(df, *, plot_titles: list = None):
+    return
+
+
+def annual_averages(df, *, plot_titles: list = None):
+    return
+
+
+def daily_variance(df, *, plot_titles: list = None):
+    return
+
+
+def flow_duration_curve(df, *, plot_titles: list = None):
+    return
