+++ geoglows/_plots/plots.py
@@ -1,0 +1,172 @@
+import matplotlib.pyplot as plt
+import pandas as pd
+import plotly.graph_objects as go
+
+from .format_tools import plotly_figure_to_html_plot
+from .matplotlib_forecasts import (
+    forecast as mpl_forecast,
+    forecast_stats as mpl_forecast_stats,
+    forecast_ensembles as mpl_forecast_ensembles
+)
+from .matplotlib_retrospective import (
+    retrospective as mpl_retrospective,
+    daily_averages as mpl_daily_averages,
+    monthly_averages as mpl_monthly_averages,
+    annual_averages as mpl_annual_averages
+)
+from .plotly_forecasts import (
+    forecast as plotly_forecast,
+    forecast_stats as plotly_forecast_stats,
+    forecast_ensembles as plotly_forecast_ensembles
+)
+from .plotly_retrospective import (
+    retrospective as plotly_retrospective,
+    daily_averages as plotly_daily_averages,
+    monthly_averages as plotly_monthly_averages,
+    annual_averages as plotly_annual_averages
+)
+
+
+def forecast(df: pd.DataFrame, *,
+             plot_type: str = 'plotly',
+             rp_df: pd.DataFrame = None,
+             plot_titles: list = None, ) -> go.Figure or plt.subplots:
+    """
+    Plots forecasted streamflow and optional return periods
+    Args:
+        df: the dataframe response from geoglows.data.forecast
+        rp_df: optional dataframe of return period data
+        plot_titles: optional list of strings to include in the figure title. each list item will be on a new line.
+
+    Returns:
+        go.Figure
+    """
+    if plot_type in ('plotly', 'html'):
+        figure = plotly_forecast(df, rp_df=rp_df, plot_titles=plot_titles)
+        if plot_type == 'html':
+            return plotly_figure_to_html_plot(figure)
+        return figure
+    elif plot_type == 'matplotlib':
+        return mpl_forecast(df, rp_df=rp_df, plot_titles=plot_titles)
+    return None
+
+
+def forecast_stats(df: pd.DataFrame, *,
+                   plot_type: str = 'plotly',
+                   rp_df: pd.DataFrame = None,
+                   plot_titles: list = None, ) -> go.Figure or plt.subplots:
+    """
+    Plots forecasted streamflow and optional return periods
+    Args:
+         df: the dataframe response from geoglows.data.forecast_stats
+         rp_df: optional dataframe of return period data
+         plot_titles: optional list of strings to include in the figure title. each list item will be on a new line.
+
+    Returns:
+         go.Figure
+    """
+    if plot_type in ('plotly', 'html'):
+        figure = plotly_forecast_stats(df, rp_df=rp_df, plot_titles=plot_titles)
+        if plot_type == 'html':
+            return plotly_figure_to_html_plot(figure)
+        return figure
+    elif plot_type in ('matplotlib', 'mpl'):
+        return mpl_forecast_stats(df, rp_df=rp_df, plot_titles=plot_titles)
+    return None
+
+
+def forecast_ensembles(df: pd.DataFrame, *,
+                       plot_type: str = 'plotly',
+                       rp_df: pd.DataFrame = None,
+                       plot_titles: list = None, ) -> go.Figure or plt.subplots:
+    """
+    Plots forecasted streamflow and optional return periods
+    Args:
+           df: the dataframe response from geoglows.data.forecast_ensembles
+           rp_df: optional dataframe of return period data
+           plot_titles: optional list of strings to include in the figure title. each list item will be on a new line.
+
+    Returns:
+           go.Figure
+    """
+    if plot_type in ('plotly', 'html'):
+        figure = plotly_forecast_ensembles(df, rp_df=rp_df, plot_titles=plot_titles)
+        if plot_type == 'html':
+            return plotly_figure_to_html_plot(figure)
+        return figure
+    elif plot_type == 'matplotlib':
+        return mpl_forecast_ensembles(df, rp_df=rp_df, plot_titles=plot_titles)
+    return None
+
+
+def retrospective(df: pd.DataFrame, *,
+                  plot_type: str = 'plotly',
+                  rp_df: pd.DataFrame = None,
+                  plot_titles: list = None, ) -> go.Figure or plt.subplots:
+    if plot_type in ('plotly', 'html'):
+        figure = plotly_retrospective(df, rp_df=rp_df, plot_titles=plot_titles)
+        if plot_type == 'html':
+            return plotly_figure_to_html_plot(figure)
+        return figure
+    if plot_type == 'matplotlib':
+        return mpl_retrospective(df, rp_df=rp_df, plot_titles=plot_titles)
+    return
+
+
+def daily_averages(df: pd.DataFrame, *,
+                   plot_type: str = 'plotly',
+                   plot_titles: list = None, ) -> go.Figure or plt.subplots:
+    if plot_type in ('plotly', 'html'):
+        figure = plotly_daily_averages(df, plot_titles=plot_titles)
+        if plot_type == 'html':
+            return plotly_figure_to_html_plot(figure)
+        return figure
+    if plot_type == 'matplotlib':
+        return mpl_daily_averages(df, plot_titles=plot_titles)
+    return
+
+
+def monthly_averages(df: pd.DataFrame, *,
+                     plot_type: str = 'plotly',
+                     plot_titles: list = None, ) -> go.Figure or plt.subplots:
+    if plot_type in ('plotly', 'html'):
+        figure = plotly_monthly_averages(df, plot_titles=plot_titles)
+        if plot_type == 'html':
+            return plotly_figure_to_html_plot(figure)
+        return figure
+    if plot_type == 'matplotlib':
+        return mpl_monthly_averages(df, plot_titles=plot_titles)
+    return
+
+
+def annual_averages(df: pd.DataFrame, *,
+                    plot_type: str = 'plotly',
+                    plot_titles: list = None, ) -> go.Figure or plt.subplots:
+    if plot_type in ('plotly', 'html'):
+        figure = plotly_annual_averages(df, plot_titles=plot_titles)
+        if plot_type == 'html':
+            return plotly_figure_to_html_plot(figure)
+        return figure
+    if plot_type == 'matplotlib':
+        return mpl_annual_averages(df, plot_titles=plot_titles)
+    return
+
+
+def daily_variance(df: pd.DataFrame, *,
+                   plot_type: str = 'plotly',
+                   plot_titles: list = None, ) -> go.Figure or plt.subplots:
+    if plot_type == 'plotly':
+        return
+    if plot_type == 'matplotlib':
+        return
+    return
+
+
+def flow_duration_curve(df: pd.DataFrame, *,
+                        plot_type: str = 'plotly',
+                        plot_titles: list = None, ) -> go.Figure or plt.subplots:
+    if plot_type == 'plotly':
+        return
+    if plot_type == 'matplotlib':
+        return
+    return
