+++ geoglows/tables.py
@@ -1,0 +1,111 @@
+import datetime
+import json
+import os
+
+import jinja2
+import pandas as pd
+
+__all__ = [
+    'probabilities_table',
+    'return_periods_table',
+]
+
+
+def probabilities_table(stats: pd.DataFrame, ensem: pd.DataFrame, rperiods: pd.DataFrame) -> str:
+    """
+    Processes the results of forecast_stats , forecast_ensembles, and return_periods and uses jinja2 template
+    rendering to generate html code that shows the probabilities of exceeding the return period flow on each day.
+
+    Args:
+        stats: the csv response from forecast_stats
+        ensem: the csv response from forecast_ensembles
+        rperiods: the csv response from return_periods
+
+    Return:
+         string containing html to build a table with a row of dates and for exceeding each return period threshold
+    """
+    dates = stats.index.tolist()
+    startdate = dates[0]
+    enddate = dates[-1]
+    span = enddate - startdate
+    uniqueday = [startdate + datetime.timedelta(days=i) for i in range(span.days + 2)]
+    # get the return periods for the stream reach
+    rp2 = rperiods['return_period_2'].values
+    rp5 = rperiods['return_period_5'].values
+    rp10 = rperiods['return_period_10'].values
+    rp25 = rperiods['return_period_25'].values
+    rp50 = rperiods['return_period_50'].values
+    rp100 = rperiods['return_period_100'].values
+    # fill the lists of things used as context in rendering the template
+    days = []
+    r2 = []
+    r5 = []
+    r10 = []
+    r25 = []
+    r50 = []
+    r100 = []
+    for i in range(len(uniqueday) - 1):  # (-1) omit the extra day used for reference only
+        tmp = ensem.loc[uniqueday[i]:uniqueday[i + 1]]
+        days.append(uniqueday[i].strftime('%b %d'))
+        num2 = 0
+        num5 = 0
+        num10 = 0
+        num25 = 0
+        num50 = 0
+        num100 = 0
+        for column in tmp:
+            column_max = tmp[column].to_numpy().max()
+            if column_max > rp100:
+                num100 += 1
+            if column_max > rp50:
+                num50 += 1
+            if column_max > rp25:
+                num25 += 1
+            if column_max > rp10:
+                num10 += 1
+            if column_max > rp5:
+                num5 += 1
+            if column_max > rp2:
+                num2 += 1
+        r2.append(round(num2 * 100 / 52))
+        r5.append(round(num5 * 100 / 52))
+        r10.append(round(num10 * 100 / 52))
+        r25.append(round(num25 * 100 / 52))
+        r50.append(round(num50 * 100 / 52))
+        r100.append(round(num100 * 100 / 52))
+    path = os.path.abspath(os.path.join(os.path.dirname(__file__), 'templates', 'probabilities_table.html'))
+    with open(path) as template:
+        return jinja2.Template(template.read()).render(days=days, r2=r2, r5=r5, r10=r10, r25=r25, r50=r50, r100=r100,
+                                                       colors=_plot_colors())
+
+
+def return_periods_table(rperiods: pd.DataFrame) -> str:
+    """
+    Processes the dataframe response from the return_periods function and uses jinja2 to render an html string for a
+    table showing each of the return periods
+
+    Args:
+        rperiods: the dataframe from the return periods function
+
+    Returns:
+        string of html
+    """
+
+    # find the correct template to render
+    path = os.path.abspath(os.path.join(os.path.dirname(__file__), 'templates', 'return_periods_table.html'))
+    # work on a copy of the dataframe so you dont ruin the original
+    tmp = rperiods
+    reach_id = str(tmp.index[0])
+    js = json.loads(tmp.to_json())
+    rp = {}
+    for i in js:
+        if i.startswith('return_period'):
+            year = i.split('_')[-1]
+            rp[f'{year} Year'] = js[i][reach_id]
+        elif i == 'max_flow':
+            rp['Max Simulated Flow'] = js[i][reach_id]
+
+    rp = {key: round(value, 2) for key, value in sorted(rp.items(), key=lambda item: item[1])}
+
+    with open(path) as template:
+        return jinja2.Template(template.read()).render(reach_id=reach_id, rp=rp, colors=_plot_colors())
